refactored code a bit, next: add checkout sheet
